requirements.txt
tensorflow
numpy
Pillow
flask
werkzeug
dominate
gunicorn